refactor: split correlation from utils
